Starting work on a new form of history navigation
We can't currently interlink WAD pages because this could cause an infinite deep bread-crumb chain with the way history navigation currently wotrks.

The plan is to not base the bread-crumb on the chain of navigation, but on a each pages particular location, e.g. "Cacowards > Cacowards 2016", maybe implied by folder heirarchy or perhaps given by the page itself, e.g. `$TITLE:=...".