dosmag: Sync tab-stops between source and screen
Definitely the very best thing I've added to dosmag! Even though the source text has formatting marks, not shown on the screen, the tab-stops on screen now align to the source file; this means, that despite the difference between the length of spans of text in the source and on the screen, you can layout your source in tidy columns and you'll get the same appearance on the screen! This gets rid of the wonky manual spacing previously needed.

If a tab only expands to a single-space then the column will not sync; this will be fixed later (I pre-convert all tabs to spaces to simplify logic, but it causes this bug).