Large reworking of launcher structure
Given the harsh limitations of QBASIC, I'm trying to do my best to make this modular. My short-term goals have shifted from building the launcher UI to getting the launcher to replicate the current "doom.bat" by supporting the same command-line parameters.

I'd like to first replace "doom.bat" with a drop-in replacement, and the launcher can then be expanded over time without having to perfect it first. Producing the full quantity of UI needed for game customisation and mod selection will take a long, long time and a lot of work and I'd rather do that piece-by-piece than all-or-nothing.

A few notes for this commit:

* Fullscreen via ALT+ENTER now uses the one screen mode rather than cycling through different stretch / anti-aliasing effects thanks to Fellippe Heitor adding my suggested `_ALLOWFULLSCREEN` command :)

* Replaced home-made `TRIM$` with new built-in `_TRIM$`

* Added registry code to read Steam / GOG paths