Start work on auto-file loading
Engines can specify a set of files to automatically load, e.g. "brightmaps.pk3;lights.pk3". We can read this data in now and loosley parse it. TODO: find the files, write to the command-line based on the engine.

This commit also updates a lot of the launcher INI files for the engines, and adds GZDoom 3.5.0 to the pile.

QB64 updtaed to fix bug with getting the screen resolution.