Be gone, thot!
FINALLY, we remove "doom.bat" & "play.bat"!

This has been a year(?) in the making. Using batch to handle intelligent engine selection was never going to be possible. Finally I have pushed the replacement launcher to the state where it can replace the old batch scripts.

There's still a handful of things to implement before we've covered all original functionality, but for now it's possible to untangle the PortaDOOM pages from "doom.bat" / "play.bat" and bring it all under "launcher.exe".

There will be a lot of 'fixing up' to follow as I go through and test the launch scripts for each game, and then ensure the build script includes all necessary files.

* Rename current GZDoom folder to "gzdoom_x86/64"; this helps with the include lists as they won't have to be updated with every new GZDoom version
* Fix engine filtering with `/USE` parameter