Rework game selection screen
We now show a title and description for the game as a whole, and descriptions for each game choice rather than titles & descriptions for the choices. This saves space and also works better for multi-choice games, particularly the test scripts for each engine. The actual INI file properties and structure may well change later on as I'm not entirely happy with the terminology used.

* Also added the INI file path as an additional search location