Fixes:
Added the ability to specify a specific GZDoom version to use, if and only if it's chosen. This can also be used to deny GZDoom by specifying a non-existant version number.

Deny GZDoom for 50 Shades of Graytall; all versions produce some kind of error in the console.

Fix SKULLDASH files not being added to the package.

Use correct version of GZDoom for Doomed Space Wars.

Update social links on home pages.