Fix IWAD discovery in launcher
This appeared super broken and worked largely by happenstance before.

When looking for the path for an IWAD, the IWAD metadata cache is queried for the path and IWAD metadata is loaded from INI if present.

IWAD names given as paths are better handled.

Support for ".iwad" & ".ipk3" IWADs added.

Getting a filename from a path was fixed for file names containing multiple dots.