Change name markup syntax
Using underscores for names has been a nuisance for a long time because it can't handle underscore affixes on names.
The syntax has been changed to angle-brackets so that the start and end are separately delineated; word boundaries are no longer required inside or outside the name.

Over 5'000 replacements were done in the pages, errors may exist and either way not every name is marked consistently.