Hedon
We're reaching about 1.5GB, on-disk, for Cacowards 2019 so Hedon and/or SRB2 might be dropped if the final compressed size is too much.

Many of the WADs have been put through DOOM-Cruncher but the large PK3s remain, including Hedon.

Hedon requires special keybinds / config so a case has been added to "make_defaults.bat" for GZDoom v3.4