Add DSDA-Doom
This supercedes PrBoom+, although PrBoom+ can still be included where required.

Keyboard codes in the config file are complex; I'll set the default controls in the next commit.

DSDA-Doom does support Heretic and Hexen IWADs but supposedly requires `-heretic` and `-hexen` switches to do so. Loading the IWADs only appears to work fine, but needs to be tested with PWADs.

Support for SIGIL's episode 5 has been added to all ports now, so the custom launch options are no longer needed.